Move FeatureFlags class into Worker namespace
