+++ tests/Functional/worker.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-use Temporal\FeatureFlags;
 use Temporal\Testing\WorkerFactory;
 use Temporal\Tests\Fixtures\PipelineProvider;
 use Temporal\Tests\Interceptor\HeaderChanger;
 use Temporal\Tests\Interceptor\InterceptorCallsCounter;
+use Temporal\Worker\FeatureFlags;
 use Temporal\Worker\WorkerInterface;
 
 require __DIR__ . '/../../vendor/autoload.php';
